fix: change TaskService based on the principle separation of concerns

+++ server/src/modules/tasks/tasks.service.ts
@@ -1,14 +1,15 @@
 import type TaskType from "./types.ts";
-import ITaskRepository from './tasks.repository.ts'
-import { isExistTaskError } from "./customErrors.ts";
+import ITaskRepository from './tasks.repository.js'
+import { isExistTaskError, ForbiddenError } from "./customErrors.js";
+import Database from "better-sqlite3";
 
 interface ITaskRepository {
     repo: {
         getTasks(user_id: number): TaskType[];
-        getTaskById(id: number, user_id: number): TaskType | undefined;
+        getTask(id: number, user_id: number): TaskType | undefined;
         createTask(task: TaskType, user_id: number): TaskType;
         updateTask(id: number, task: TaskType, user_id: number): TaskType;
-        deleteTask(id: number, user_id: number): boolean | undefined;
+        deleteTask(id: number, user_id: number): Database.RunResult;
     }
 }
 
@@ -19,11 +20,18 @@ function TaskService({ repo }: ITaskRepository) {
         return tasks
     }
 
-    const getTask = (id: number, user_id: number) => {
-        const task = repo.getTaskById(id, user_id);
+    const getTask = (task_id: number, user_id: number) => {
+
+        const task = repo.getTask(task_id, user_id);
+
         if (!task) {
-            throw new isExistTaskError(`no task with ${id}`);
+            throw new isExistTaskError(`no task with ${task_id}`, 404);
         }
+
+        if (task.user_id !== user_id) {
+            throw new ForbiddenError('access denied', 403);
+        }
+
         return task
     }
 
@@ -33,21 +41,36 @@ function TaskService({ repo }: ITaskRepository) {
     }
 
     const updateTask = (id: number, task: TaskType, user_id: number) => {
-        const existedTask = repo.getTaskById(id, user_id);
+        const existedTask = repo.getTask(id, user_id);
+
         if (!existedTask) {
-            throw new isExistTaskError(`no task with ${id}`);
+            throw new isExistTaskError(`no task with ${id}`, 404);
         }
+
+        if (existedTask.user_id !== user_id) {
+            throw new ForbiddenError('access denied', 403)
+        }
+
         const updatedTask = repo.updateTask(id, task, user_id);
         return updatedTask
     }
 
     const deleteTask = (id: number, user_id: number) => {
-        const existedTask = repo.getTaskById(id, user_id);
+        const existedTask = repo.getTask(id, user_id);
+
         if (!existedTask) {
-            throw new isExistTaskError(`no task with ${id}`);
+            throw new isExistTaskError(`no task with ${id}`, 404);
         }
-        repo.deleteTask(id, user_id);
-        return true
+
+        if (existedTask.user_id !== user_id) {
+            throw new ForbiddenError('access denied', 403)
+        }
+
+        const result = repo.deleteTask(id, user_id);
+        if (result) {
+            throw new isExistTaskError(`no task with ${id}`, 404);
+        }
+        return;
     }
 
     return { getTasks, getTask, createTask, updateTask, deleteTask }
